chore(providers): update kv-nats to prefer secrets
