+++ crates/provider-keyvalue-nats/README.md
@@ -17,12 +17,18 @@ This provider is multi-threaded and can handle concurrent requests from multiple
 To configure this provider, use the following settings in link definitions:
 
 | **Property**  | **Description**                                                                                                                                                                                                                                                                                                 |
-| :------------ | :-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+|:--------------|:----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | `CLUSTER_URI` | NATS cluster connection URI. If not specified, the default is `nats://0.0.0.0:4222`                                                                                                                                                                                                                             |
 | `JS_DOMAIN`   | Optional NATS Jetstream domain to connect to.                                                                                                                                                                                                                                                                   |
 | `BUCKET`      | The name of an existing NATS Kv Store. It must be specified. Additional links could be added if access to more Kv stores is needed; the buckets could be referenced by their respective `link_names` (please see the Rust ***keyvalue-messaging*** example for a comprehensive demonstration of this approach). |
-| `CLIENT_JWT`  | Optional JWT auth token. For JWT authentication, both `CLIENT_JWT` and `CLIENT_SEED` must be provided.                                                                                                                                                                                                          |
-| `CLIENT_SEED` | Private seed for JWT authentication.                                                                                                                                                                                                                                                                            |
-| `TLS_CA`      | To secure communications with the NATS server, the public key of its CA could be provided as an encoded string.                                                                                                                                                                                                 |
-| `TLS_CA_FILE` | Alternatively, the path qualified name of the CA public key could be provided. If both are provided, the TLS\_CA will be used.                                                                                                                                                                                  |
+| `TLS_CA_FILE` | Alternatively, the path qualified name of the CA public key could be provided. If both are provided, the `TLS_CA` will be used.                                                                                                                                                                                 |
 
+## Link Definition Secret Settings
+
+While the provider supports receiving the following values via configuration (similar to values outlined in the configuration section above), the values below are *sensitive*, and thus *should* be configured via link-time secrets.
+
+| **Property**  | **Description**                                                                                                 |
+|:--------------|:----------------------------------------------------------------------------------------------------------------|
+| `CLIENT_JWT`  | Optional JWT auth token. For JWT authentication, both `CLIENT_JWT` and `CLIENT_SEED` must be provided.          |
+| `CLIENT_SEED` | Private seed for JWT authentication.                                                                            |
+| `TLS_CA`      | To secure communications with the NATS server, the public key of its CA could be provided as an encoded string. |
